+++ addon/content/components/conversation/conversationFooter.jsx
@@ -5,7 +5,7 @@
 import React from "react";
 import * as ReactRedux from "react-redux";
 import PropTypes from "prop-types";
-import { summaryActions } from "./reducer-summary.js";
+import { summaryActions } from "../../reducer/reducer-summary.js";
 
 class _ConversationFooter extends React.PureComponent {
   constructor(props) {
