+++ addon/content/components/conversation/conversationHeader.jsx
@@ -5,9 +5,9 @@
 import React from "react";
 import * as ReactRedux from "react-redux";
 import PropTypes from "prop-types";
-import { messageActions } from "./reducer-messages.js";
-import { summaryActions } from "./reducer-summary.js";
-import { SvgIcon } from "./svgIcon.jsx";
+import { messageActions } from "../../reducer/reducer-messages.js";
+import { summaryActions } from "../../reducer/reducer-summary.js";
+import { SvgIcon } from "../svgIcon.jsx";
 
 const LINKS_REGEX = /((\w+):\/\/[^<>()'"\s]+|www(\.[-\w]+){2,})/;
