+++ addon/content/components/message/messageTags.jsx
@@ -4,8 +4,7 @@
 
 import React from "react";
 import PropTypes from "prop-types";
-import { SvgIcon } from "./svgIcon.jsx";
-
+import { SvgIcon } from "../svgIcon.jsx";
 /**
  * Determine if a background color is light enough to require dark text.
  *
