+++ addon/content/components/svgIcon.jsx
