+++ addon/content/reducer/reducer-attachments.js
